+++ src/views/KeyResultHome.vue
@@ -5,12 +5,12 @@
       <p>{{ activeKeyResult.description }}</p>
 
       <div class="main-widgets">
-        <div class="main-widgets__current main-widgets__current__children">
+        <div class="main-widgets__current main-widgets__current--children">
           <h3 class="main-widgets__title">
             <i class="fas fa-chart-line" />
             {{ $t('keyres.registerProgression.value') }}
           </h3>
-          <div class="main-widgets__current__value">
+          <div class="main-widgets__current--value">
             {{
               typeof activeKeyResult.currentValue === 'undefined'
                 ? activeKeyResult.startValue
@@ -18,7 +18,7 @@
             }}
           </div>
 
-          <div class="main-widgets__current__unit">
+          <div class="main-widgets__current--unit">
             {{ activeKeyResult.unit }}
           </div>
 
@@ -226,98 +226,6 @@ export default {
 .keyres {
   display: flex;
   flex-wrap: wrap;
-}
-
-.main-widgets {
-  display: flex;
-  flex-direction: row;
-  width: 100%;
-  margin-top: 1.5rem;
-  margin-bottom: 2rem;
-
-  @media screen and (max-width: bp(s)) {
-    flex-direction: column;
-  }
-
-  &__title {
-    margin-bottom: 0.5rem;
-    color: $color-grey-400;
-    font-weight: 500;
-    font-size: $font-size-0;
-    text-transform: uppercase;
-  }
-
-  &__current {
-    align-self: flex-start;
-    width: span(3);
-    margin-bottom: 0.5rem;
-
-    padding: 1rem;
-    background: white;
-    box-shadow: 0 2px 3px rgba(black, 0.1);
-
-    @media screen and (max-width: bp(s)) {
-      width: span(9);
-    }
-
-    @media screen and (min-width: bp(m)) {
-      width: span(2, 0, span(9));
-    }
-
-    @media screen and (min-width: bp(l)) {
-      width: span(2, 0, span(7));
-    }
-
-    @media screen and (min-width: bp(xl)) {
-      width: span(2, 0, span(6));
-    }
-
-    &__value {
-      color: $color-yellow;
-      font-weight: 700;
-      font-size: 50px;
-    }
-
-    &__unit {
-      margin-top: -0.5rem;
-      margin-bottom: 0.5rem;
-      font-weight: 500;
-      text-transform: capitalize;
-    }
-
-    &__children {
-      display: flex;
-      flex-direction: column;
-    }
-  }
-
-  &__graph {
-    width: span(9);
-
-    margin-left: span(0, 1);
-    padding: 1rem;
-    background: white;
-    box-shadow: 0 2px 3px rgba(black, 0.1);
-
-    @media screen and (max-width: bp(s)) {
-      margin-left: span(0);
-    }
-
-    @media screen and (min-width: bp(m)) {
-      width: span(7, 0, span(9));
-      margin-left: span(0, 1, span(9));
-    }
-
-    @media screen and (min-width: bp(l)) {
-      width: span(5, 0, span(7));
-      margin-left: span(0, 1, span(7));
-    }
-
-    @media screen and (min-width: bp(xl)) {
-      width: span(4, 0, span(6));
-      margin-left: span(0, 1, span(6));
-    }
-  }
 }
 
 .main__table {